[REF] General refactoring/fixing for package engine.builders
